remove unused parameter to pack all subs, moved creation of WApplication into a helper method that can be called from unit tests, and fixed other minor warnings

+++ trunk/fix/users.cpp
@@ -60,7 +60,7 @@ int FixUsersCommand::Execute() {
 		if (!user.IsUserDeleted() && !user.IsUserInactive()) {
 			smalrec sr = { 0 };
 			strcpy((char*) sr.name, user.GetName());
-			sr.number = i;
+			sr.number = static_cast<unsigned short>(i);
 			std::string namestring((char*) sr.name);
 			if (names.find(namestring) == names.end()) {
 				smallrecords.push_back(sr);
